requirements.txt
PyYAML==6.0.1
vosk==0.3.45
piper-tts==1.2.0
pyaudio==0.2.13
SpeechRecognition==3.10.0
dbus-python==1.3.2
pulsectl==24.12.0
smbus2==0.6.0
gpiozero==2.0
psutil==7.2.0
# numpy и pillow ставим через apt (python3-numpy), здесь они для справки